+++ freetokenapi/__init__.py
@@ -2,6 +2,6 @@ from __future__ import annotations
 
 from .logging import configure
 
-__version__ = "2.0.0"
+__version__ = "2.0.1"
 
 configure()